requirements.txt
django>=5.0
djangorestframework
drf-spectacular
django-environ
django-cors-headers
pymysql
cryptography
django-redis
celery
redis
pillow
requests
django-allauth
django-tailwind
python-dotenv
easyocr
pyzbar
opencv-python-headless
ollama
django-allauth[socialaccount]
vt-py
channels
channels-redis
daphne
beautifulsoup4
django-otp
pyotp
qrcode
django-turnstile
